create: likeeee

+++ Backend/Model/Person.cs
@@ -4,12 +4,12 @@ public class PersonData
 {
     public int Gender { get; set; }
     public int Age { get; set; }
-    public int SleepDuration { get; set; }
-    public int Physical { get; set; }
+    public float SleepDuration { get; set; }
+    public float Physical { get; set; }
     public int StressLevel { get; set; }
     public int BMI { get; set; }
-    public int BloodPressure { get; set; }
-    public int Heart { get; set; }
+    public float BloodPressure { get; set; }
+    public float Heart { get; set; }
     public int Steps { get; set; }
     public int SleepDisorder { get; set; }
 }
